+++ src/controllers/api/user/index.ts
@@ -1,0 +1,1 @@
+export { UserController } from './class';
