+++ src/controllers/index.ts
@@ -1,2 +1,2 @@
+export { UserController } from './api';
 export { AuthMiddleware, IAuthMiddleware } from './middleware';
-export { UserController } from './user';
